Add download operations
The ability to upload images has been added to blogpost, aboutme and project controllers. File.api has been updated and now, if there is a file with the same file name, it adds a phrase (1) next to it. The number of the file works as +1. A slider entity has been created for the slider item on the main page. Features have been added so that 2 files can be loaded into it.